rename unix -> system for more accurate file struct representation
